fetch/metadata/generated/appcache-manifest.https.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/appcache-manifest.sub.https.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for Appcache manifest</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url) {
    const iframe = document.createElement('iframe');
    iframe.src =
      '/fetch/metadata/resources/appcache-iframe.sub.html?manifest=' + encodeURIComponent(url);

    return new Promise((resolve) => {
        addEventListener('message', function onMessage(event) {
          if (event.source !== iframe.contentWindow) {
            return;
          }
          removeEventListener('message', onMessage);
          resolve(event.data);
        });

        document.body.appendChild(iframe);
      })
      .then((message) => {
        if (message !== 'okay') {
          throw message;
        }
      })
      .then(() => iframe.remove());
  }

  async_test((t) => {
    const key = '{{uuid()}}';
    assert_implements_optional(
      !!window.applicationCache, 'Application Cache supported.'
    );

    induceRequest(makeRequestURL(key, ['httpsOrigin']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-origin']);
      })
      .then(() => t.done(), t.step_func((error) => { throw error; }));
  }, 'sec-fetch-frame-ancestors - Same origin');

  async_test((t) => {
    const key = '{{uuid()}}';
    assert_implements_optional(
      !!window.applicationCache, 'Application Cache supported.'
    );

    induceRequest(makeRequestURL(key, ['httpsCrossSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['cross-site']);
      })
      .then(() => t.done(), t.step_func((error) => { throw error; }));
  }, 'sec-fetch-frame-ancestors - Cross-site');

  async_test((t) => {
    const key = '{{uuid()}}';
    assert_implements_optional(
      !!window.applicationCache, 'Application Cache supported.'
    );

    induceRequest(makeRequestURL(key, ['httpsSameSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-site']);
      })
      .then(() => t.done(), t.step_func((error) => { throw error; }));
  }, 'sec-fetch-frame-ancestors - Same site');
  </script>
  </body>
</html>
